refactor: move browser cores to ESM
