+++ AGENTS.md
@@ -126,7 +126,7 @@
   interface.
 - Harnessed artifact-generation flows now proactively refresh the owner-module
   mtimes for embedded package assets such as `open-target-details.js`,
-  `blueprint-graph-core.js`, `blueprint-preview-core.js`,
+  `blueprint-graph-core.mjs`, `blueprint-preview-core.mjs`,
   `preview-runtime-base.js`, `preview-runtime-data.js`,
   `preview-runtime-render.js`, `preview-runtime-hydration.js`,
   `preview-runtime-lifecycle.js`, `preview-runtime-surface.js`,
