+++ doc/DESIGN_RATIONALE.md
@@ -62,9 +62,9 @@ Command modules are split by concern:
 - generic command CSS and shared preview/runtime asset injection in
   `VersoBlueprint/Commands/Common.lean`
 - the target-details opener in `VersoBlueprint/Commands/open-target-details.js`
-- shared graph-data browser helpers in `VersoBlueprint/blueprint-graph-core.js`
+- shared graph-data browser helpers in `VersoBlueprint/blueprint-graph-core.mjs`
 - shared preview URL/key browser helpers in
-  `VersoBlueprint/blueprint-preview-core.js`
+  `VersoBlueprint/blueprint-preview-core.mjs`
 - the private preview base/debug helpers in
   `VersoBlueprint/Commands/preview-runtime-base.js`
 - the private preview-data/cache runtime helpers in
@@ -126,8 +126,8 @@ Shared and feature-specific browser assets stay with their owning commands:
 - `Commands/open-target-details.js`
 - `Commands/graph-runtime-core.js`
 - `Commands/graph.js`
-- `blueprint-graph-core.js`
-- `blueprint-preview-core.js`
+- `blueprint-graph-core.mjs`
+- `blueprint-preview-core.mjs`
 - `Informal/Block/relation-panel.js`
 
 Per-command CSS overlays stay with their commands:
@@ -366,8 +366,8 @@ hydration.
 | External declaration rows and rendered declaration body strategy | `Informal.ExternalCode.renderExternalDeclRowsWith` | external-code panels and HTML-cache-backed Lean-code previews | shared row/status/footer owner with page-hover and self-contained body strategies |
 | Manifest/cache-backed block shell assembly | `Informal.PreviewManifest.BlockRender.renderWithRenderedContent` | Slides, grafts, and custom generated consumers | single manifest-backed assembly owner; callers only supply config and content |
 | Graft node lookup, diagnostics, and outer graft attrs | `Informal.Graft.renderNodeFromManifestCache` / `renderNodeWithContent` | Manual grafts, Slides grafts, external generated consumers | single graft owner |
-| Browser generated-data URLs and preview keys | `blueprint-preview-core.js` shared by the bundled runtime and `api/preview.mjs` | browser runtime, ESM preview clients, custom generated pages | single JS owner for generated-data URL construction and preview-key formatting; runtime stores and ESM helpers delegate to it instead of reimplementing the same string rules |
-| Browser graph JSON discovery and graph manifest loading | `blueprint-graph-core.js` shared by the bundled runtime and `api/graph.mjs` | graph command runtime, graph dashboards, custom browser clients | single JS owner for graph-data normalization and graph JSON/script discovery; compatibility globals are exports, not the bundled runtime's internal data path |
+| Browser generated-data URLs and preview keys | `blueprint-preview-core.mjs` shared by the bundled runtime and `api/preview.mjs` | browser runtime, ESM preview clients, custom generated pages | single ESM owner for generated-data URL construction and preview-key formatting; runtime stores and ESM helpers delegate to it instead of reimplementing the same string rules, and current Verso page assets use a generated classic-script adapter |
+| Browser graph JSON discovery and graph manifest loading | `blueprint-graph-core.mjs` shared by the bundled runtime and `api/graph.mjs` | graph command runtime, graph dashboards, custom browser clients | single ESM owner for graph-data normalization and graph JSON/script discovery; compatibility globals are installed by the adapter, not owned by the bundled runtime's internal data path |
 | Browser manifest/cache loading and body-fragment insertion | `Commands/preview-runtime-data.js` for data/cache loading plus `Commands/preview-runtime-render.js` `resolvePreview` and `renderPreviewInto`, with `Commands/preview-runtime-surface.js` `resolvePreviewHtml` and `renderPreviewIntoSurface` adapting those helpers for bundled surfaces | graph, summary, relation panels, inline previews, custom browser clients | single JS data/cache store owner; the data chunk delegates URL/key primitives to preview core and graph data to graph core, while the render chunk joins those entries with rendered fragments |
 | Browser canonical generated-node insertion | `Commands/preview-runtime-render.js` `resolveCanonicalPreview` and `renderCanonicalPreviewInto` | standalone/custom browser clients that want regular Blueprint node visuals | single JS canonical-preview owner |
 | Browser preview panel behavior | `Commands/preview-runtime-surface.js` `createPreviewSurface`, `hidePreviewSurfaces`, and panel helpers plus `Commands/preview-runtime-lifecycle.js` trigger/dismiss/reposition helpers | summary, code-summary, inline-preview, relation-panel, Slides, and graph feature scripts | single JS behavior helper; feature scripts pass selectors/defaults through rendered descriptors or feature-specific callbacks instead of owning panel slots, trigger lifetimes, dismissal binding, or close-button wiring |
@@ -444,7 +444,7 @@ The workflow implies a few constraints for renderers:
 
 - **Split JavaScript by responsibility, not feature semantics.**
   The current preview runtime is still emitted as one bundled asset. Its source
-  is now split across private source chunks: `blueprint-preview-core.js` owns
+  is now split across private source chunks: `blueprint-preview-core.mjs` owns
   generated-data URL helpers and preview-key construction shared by bundled
   pages and `api/preview.mjs`; `preview-runtime-data.js` owns manifest/cache
   decoding and loading, load-status readers, entry lookup, and graph-core
@@ -469,7 +469,7 @@ The workflow implies a few constraints for renderers:
   | Boundary | Current responsibility | Split target |
   | --- | --- | --- |
   | API readiness | Install `window.VersoBlueprint`, queue `onRenderReady` callbacks, and expose the stable render API. | `Commands/preview-runtime.js` source chunk inside the bundled runtime |
-  | Preview URL/key primitives | Resolve `-verso-data` URLs and normalize preview keys for both bundled and ESM clients. | `blueprint-preview-core.js` shared implementation file |
+  | Preview URL/key primitives | Resolve `-verso-data` URLs and normalize preview keys for both bundled and ESM clients. | `blueprint-preview-core.mjs` shared implementation file, embedded through a generated classic adapter for current Verso output |
   | Preview data access | Load manifest/cache JSON, keep load status, delegate graph data to graph core, and look up entries. | `Commands/preview-runtime-data.js` source chunk inside the bundled runtime |
   | Fragment rendering | Resolve manifest/cache pairs, produce diagnostics, insert rendered fragments, and fetch canonical node wrappers. | `Commands/preview-runtime-render.js` source chunk inside the bundled runtime |
   | Hydration registry | Run math rendering and feature hydrators after custom insertion. | `Commands/preview-runtime-hydration.js` source chunk inside the bundled runtime |
