+++ src/VersoBlueprint/PreviewManifest.lean
@@ -549,9 +549,9 @@ def htmlCacheFilename : String := "blueprint-html-cache.json"
 
 def graphApiModuleFilename : String := "blueprint-graph-api.mjs"
 
-def graphCoreModuleFilename : String := "blueprint-graph-core.js"
+def graphCoreModuleFilename : String := "blueprint-graph-core.mjs"
 
-def previewCoreModuleFilename : String := "blueprint-preview-core.js"
+def previewCoreModuleFilename : String := "blueprint-preview-core.mjs"
 
 def previewApiModuleFilename : String := "blueprint-preview-api.mjs"
 
@@ -566,9 +566,9 @@ def graphApiModulePath : String := apiModuleDirname ++ "/" ++ graphApiModuleAlia
 def previewApiModulePath : String := apiModuleDirname ++ "/" ++ previewApiModuleAliasFilename
 
 -- Keep this module rebuilt when the standalone browser ESM APIs change.
-private def graphCoreModuleJs : String := include_str "blueprint-graph-core.js"
+private def graphCoreModuleMjs : String := include_str "blueprint-graph-core.mjs"
 
-private def previewCoreModuleJs : String := include_str "blueprint-preview-core.js"
+private def previewCoreModuleMjs : String := include_str "blueprint-preview-core.mjs"
 
 private def graphApiModuleJs : String := include_str "blueprint-graph-api.mjs"
 
@@ -1663,8 +1663,8 @@ def emitBlueprintPreviewData (extensionImpls : ExtensionImpls) : ExtraStep := fu
   IO.FS.createDirAll apiDir
   IO.FS.writeFile (dataDir / manifestFilename) (toJson files.manifest).compress
   IO.FS.writeFile (dataDir / htmlCacheFilename) (toJson files.htmlCache).compress
-  IO.FS.writeFile (dataDir / graphCoreModuleFilename) graphCoreModuleJs
-  IO.FS.writeFile (dataDir / previewCoreModuleFilename) previewCoreModuleJs
+  IO.FS.writeFile (dataDir / graphCoreModuleFilename) graphCoreModuleMjs
+  IO.FS.writeFile (dataDir / previewCoreModuleFilename) previewCoreModuleMjs
   IO.FS.writeFile (dataDir / graphApiModuleFilename) graphApiModuleJs
   IO.FS.writeFile (dataDir / previewApiModuleFilename) previewApiModuleJs
   IO.FS.writeFile (apiDir / graphApiModuleAliasFilename) graphApiModuleAliasJs
